feedback: rename connectHitsInsights to connectHitInsights
