+++ packages/react-instantsearch-core/src/connectors/connectHitInsights.ts
