+++ packages/react-instantsearch-core/src/index.js
@@ -48,6 +48,4 @@ export { default as connectStats } from './connectors/connectStats';
 export {
   default as connectToggleRefinement,
 } from './connectors/connectToggleRefinement';
-export {
-  default as connectHitsInsights,
-} from './connectors/connectHitsInsights';
+export { default as connectHitInsights } from './connectors/connectHitInsights';
